+++ packages/snaps-controllers/src/interface/SnapInterfaceController.ts
@@ -5,8 +5,8 @@ import type {
 import type {
   ControllerGetStateAction,
   ControllerStateChangeEvent,
-} from '@metamask/base-controller/next';
-import { BaseController } from '@metamask/base-controller/next';
+} from '@metamask/base-controller';
+import { BaseController } from '@metamask/base-controller';
 import type { Messenger } from '@metamask/messenger';
 import type { TestOrigin } from '@metamask/phishing-controller';
 import type {