+++ packages/snaps-controllers/src/snaps/SnapController.test.tsx
@@ -1,5 +1,7 @@
-import { getPersistentState } from '@metamask/base-controller';
-import { deriveStateFromMetadata } from '@metamask/base-controller/next';
+import {
+  getPersistentState,
+  deriveStateFromMetadata,
+} from '@metamask/base-controller';
 import { encrypt } from '@metamask/browser-passworder';
 import {
   createAsyncMiddleware,