+++ packages/snaps-controllers/src/snaps/SnapController.ts
@@ -6,8 +6,8 @@ import {
 import type {
   ControllerGetStateAction,
   ControllerStateChangeEvent,
-} from '@metamask/base-controller/next';
-import { BaseController } from '@metamask/base-controller/next';
+} from '@metamask/base-controller';
+import { BaseController } from '@metamask/base-controller';
 import type { CryptographicFunctions } from '@metamask/key-tree';
 import type { Messenger } from '@metamask/messenger';
 import type {